servo profiling bingchilling?

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Hardware/util/Timer.java
@@ -8,7 +8,7 @@ public class Timer {
         this.opMode = opMode;
     }
 
-    public void safeDelay(long delay){
+    public void safeDelay(double delay){
         long start = System.currentTimeMillis();
         while((System.currentTimeMillis() - start < delay) && opMode.opModeIsActive()){
             //wait
